requirements.txt
uvicorn
fastapi
psycopg2-binary
python-multipart
jinja2
faker
